Kill stale pidfiles before starting the service
If you kill postgres badly, it leaves a `postmaster.pid` file lying
around which prevents the service from starting later. This gets in the
way of development, causes confusion, and *burns battery* as launchd
tries again and again to start postgres.

There are potential consequences to this in the case of a corrupt
database, but they're so unlikely that I much prefer to err on the side
of convenience for a developer machine.